server/pyproject.toml
[tool.poetry]
name = "portal-backend"
version = "1.0.0"
description = "Django backend for the core portal."
authors = ["TACC-WMA <wma-portals@tacc.utexas.edu>"]

[tool.poetry.dependencies]
Django = "^2.2.28"
python = "^3.7"
elasticsearch-dsl = "^7.2.1"
django-bootstrap4 = "^2.0.1"
paramiko = "^2.10.1"
rt = "^2.0.0"
sqlparse = "^0.3.1"
psycopg2 = "2.8.5"
django-settings-export = "^1.2.1"
redis = "^3.5.3"
celery = "^5.2.2"
django-termsandconditions = "^2.0.6"
certifi = "^2020.4.5"
PyJWT = "^1.7.1"
kombu = "^5.2.3"
cryptography = "^3.3"
agavepy = "^v1.0.0a12"
pytz = "^2020.1"
cached-property = "^1.5.1"
ipython = "^7.16.3"
pyOpenSSL = "^19.1.0"
pycryptodome = "^3.9.7"
elasticsearch = "^7.7.1"
uwsgi = "^2.0.18"
requests = "^2.23.0"
django-impersonate = "^1.5"
channels = "^2.4.0"
channels-redis = "^3.1.0"
pytest-asyncio = "^0.14.0"
twisted = "^22.4.0"
pytas = {git = "https://bitbucket.org/taccaci/pytas.git", tag = "v1.6.0"}
gevent = "^20.9.0"
google-auth = "^1.23.0"
google-auth-oauthlib = "^0.4.2"
google-api-python-client = "^1.12.5"
jsonpickle = "^1.4.1"
python-magic = "^0.4.18"
zeep = "^4.1.0"
tapipy = "^1.2.10"

[tool.poetry.dev-dependencies]
mock = "^4.0.2"
pytest-mock = "^3.1.1"
pytest = "^5.4.3"
pytest-cov = "^2.9.0"
pytest-django = "^3.9.0"
flake8 = "^3.8.2"
coverage = "^5.1"
requests-mock = "^1.8.0"

[build-system]
requires = ["poetry>=1.1.0"]
build-backend = "poetry.masonry.api"
